requirements.txt
requests>=2.25.1
Pillow>=8.3.2
numpy>=1.21.0
torchvision